add Rest API

+++ src/main/java/com/example/project_for_nadya/controller/MyRestController.java
@@ -1,72 +1,55 @@
 package com.example.project_for_nadya.controller;
 
 import com.example.project_for_nadya.entity.Child;
+import com.example.project_for_nadya.exeption_handling.NoSuchChildExeption;
 import com.example.project_for_nadya.service.ReportService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@Controller
+@RestController
+@RequestMapping("/data")
 public class MyRestController {
 
     @Autowired
     private ReportService reportService;
 
-
-    @RequestMapping("/")
-    public String showAllChild(Model model) {
-
-        List<Child> allChild = reportService.getAllChild();
-        model.addAttribute("allChild", allChild);
-        return "all-Child";
+    @GetMapping("/child")
+    public List<Child> showAllChildren() {
+        List<Child> allChildren = reportService.getAllChildren();
+        return allChildren;
     }
 
-    @RequestMapping("/showAllDebtor")
-    public String showAllDebtor(Model model) {
-        List<Child> allDebtor = reportService.getAllDebtor();
-        model.addAttribute("allDebtor", allDebtor);
-        return "all-Debtor";
-    }
-
-
-    @RequestMapping("/addNewChild")
-//    @PreAuthorize("hasRole('ROLE_EMPLOYEE') OR hasRole('ROLE_HR')")
-    public String addNewChild(Model model) {
-        Child child = new Child();
-        model.addAttribute("child", child);
-        return "child-info";
-    }
-
-    @RequestMapping("/saveChild")
-    public String saveChild(@ModelAttribute("child") Child child) {
-        reportService.saveChild(child);
-        return "redirect:/";
-    }
-
-    @RequestMapping("/mainPage")
-    public String mainPage() {
-        return "redirect:/";
-    }
-
-
-    @RequestMapping("/updateInfo")
-//    @PreAuthorize("hasRole('ROLE_MANAGER')")
-    public String updateChild(@RequestParam("childId") int id, Model model) {
+    @GetMapping("/child/{id}")
+    public Child getChild(@PathVariable int id) {
         Child child = reportService.getChild(id);
-        model.addAttribute("child", child);
-        return "child-info";
+        if (child == null) {
+            throw new NoSuchChildExeption("There is no Child with" + id + " in our Database");
+        }
+        return child;
     }
 
+    @PostMapping("/child")
+    public Child addNewChild(@RequestBody Child child) {
+        reportService.saveChild(child);
+        return child;
+    }
 
-    @RequestMapping("/deleteChild")
-//    @PreAuthorize("hasRole('ROLE_MANAGER')")
-    public String deleteChild(@RequestParam("childId") int id) {
+    @PutMapping("/child")
+    public Child updateChild(@RequestBody Child child) {
+        reportService.saveChild(child);
+        return child;
+    }
+
+    @DeleteMapping("/child/{id}")
+    public String deleteChild(@PathVariable int id) {
+        Child child = reportService.getChild(id);
+        if (child == null) {
+            throw new NoSuchChildExeption("There is no Child with" + id + " in our Database");
+        }
         reportService.deleteChild(id);
-        return "redirect:/";
+        return "Child with ID =" + id + " was deleted";
     }
